fix: consolidate and improve TypeScript type definitions
- Consolidate all type definitions into a single file (app/types/window.d.ts)
- Add proper type definitions for all external libraries:
  - PDF.js (pdfjsLib)
  - Tesseract.js
  - jsPDF
  - PDFLib
- Make all interfaces exportable
- Add proper window augmentation
- Remove redundant type definition files
- Fix naming consistency and add missing types

This commit resolves TypeScript errors related to missing type definitions
and improves type safety across the application.
